pyproject.toml
[tool.poetry]
name = "plausibl"
version = "0.1.0"
description = ""
authors = ["hooray-sandpang <kemutakejosui@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7, <3.8"
torchvision = "0.9.1"
pyyaml = "^6.0.1"
tensorboard = "2.11.2"
scipy = "1.7.3"
pillow = "9.5"
imageio = "2.31.2"
ninja = "^1.11.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
